Dependencies Upgrade : 1) Upgraded spring dependencies to latest versions. 2) Upgraded apache commons and hikaricp dependencies. 3) Refactored code to remove deprecated code after upgrading to latest versions.

+++ wavemaker-app-runtime-core/src/main/java/com/wavemaker/runtime/data/dao/query/providers/RuntimeQueryProvider.java
@@ -78,7 +78,7 @@ public class RuntimeQueryProvider<R> implements QueryProvider<R>, PaginatedQuery
         Query<R> hibernateQuery = createQuery(session, sortedQuery, responseType);
 
         if (pageable != null) {
-            hibernateQuery.setFirstResult(pageable.getOffset());
+            hibernateQuery.setFirstResult((int) pageable.getOffset());
             hibernateQuery.setMaxResults(pageable.getPageSize());
         }
