+++ wavemaker-app-runtime-core/src/main/java/com/wavemaker/runtime/data/dao/query/providers/SessionBackedQueryProvider.java
@@ -50,7 +50,7 @@ public class SessionBackedQueryProvider<R> implements QueryProvider<R>, Paginate
             query = getAndConfigureQuery(session, name, responseType);
         }
 
-        query.setFirstResult(pageable.getOffset())
+        query.setFirstResult((int) pageable.getOffset())
                 .setMaxResults(pageable.getPageSize());
 
         return query;
